# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Yastar API - multi-module business decision software
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calculate/reverse-target:
    post:
      operationId: calculateReverseTarget
      tags: [calculate]
      summary: Reverse-calculate clients needed to hit a target profit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReverseTargetInput"
      responses:
        "200":
          description: Calculated result with feasibility check and insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReverseTargetResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculate/break-even:
    post:
      operationId: calculateBreakEven
      tags: [calculate]
      summary: Calculate how long/how many clients a new hire needs to break even
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakEvenInput"
      responses:
        "200":
          description: Break-even projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakEvenResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculate/hpp:
    post:
      operationId: calculateHpp
      tags: [calculate]
      summary: Calculate HPP (Harga Pokok Produksi/Layanan) for a product or service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HppInput"
      responses:
        "200":
          description: HPP calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HppResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculate/bep-usaha:
    post:
      operationId: calculateBepUsaha
      tags: [calculate]
      summary: Calculate business-level Break-Even Point (BEP Usaha)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BepUsahaInput"
      responses:
        "200":
          description: BEP Usaha result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BepUsahaResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculate/harga-jual:
    post:
      operationId: calculateHargaJual
      tags: [calculate]
      summary: Calculate optimal selling price or actual margin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HargaJualInput"
      responses:
        "200":
          description: Harga jual result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HargaJualResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculate/pajak:
    post:
      operationId: calculatePajak
      tags: [calculate]
      summary: Estimate monthly tax for UMKM (PPh Final 0.5% or PPh Pasal 17)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PajakInput"
      responses:
        "200":
          description: Pajak estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PajakResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculate/ekspansi:
    post:
      operationId: calculateEkspansi
      tags: [calculate]
      summary: Evaluate feasibility of opening a new branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EkspansiInput"
      responses:
        "200":
          description: Ekspansi feasibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EkspansiResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calculate/pinjaman:
    post:
      operationId: calculatePinjaman
      tags: [calculate]
      summary: Simulate a business loan (flat or annuity)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinjamanInput"
      responses:
        "200":
          description: Pinjaman simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinjamanResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/login:
    post:
      operationId: loginAdmin
      tags: [admin]
      summary: Log into the internal admin dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Check if the current admin session is valid
      responses:
        "200":
          description: Admin session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/accounts:
    post:
      operationId: createAdminAccount
      tags: [admin]
      summary: Create a new owner account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateAccountInput"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listAdminAccounts
      tags: [admin]
      summary: Search owner accounts by email (admin only)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matching accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAccountSummary"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/accounts/{id}:
    get:
      operationId: getAdminAccount
      tags: [admin]
      summary: Get an owner account with its change history (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountDetail"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminAccount
      tags: [admin]
      summary: Set an owner's tier, scenario limit, module access, and package expiry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    BusinessType:
      type: string
      enum: [barbershop, salon, nail, spa, custom]
    CommissionModel:
      type: string
      enum: [flat, base_plus_commission, tiered]
...
    ReverseTargetInput:
      type: object
      required:
        - businessType
        - employeeCount
        - workingDaysPerMonth
        - workingHoursPerDay
        - fixedCosts
        - targetProfit
        - commissionModel
        - commissionConfig
        - services
      properties:
        businessType:
          $ref: "#/components/schemas/BusinessType"
        employeeCount:
          type: integer
          minimum: 1
        workingDaysPerMonth:
          type: integer
          minimum: 1
          maximum: 31
        workingHoursPerDay:
          type: number
          minimum: 1
          maximum: 24
        fixedCosts:
          type: number
          minimum: 0
        targetProfit:
          type: number
          minimum: 0
        commissionModel:
          $ref: "#/components/schemas/CommissionModel"
        commissionConfig:
          $ref: "#/components/schemas/CommissionConfig"
        services:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceItem"
    ReverseTargetResult:
      type: object
      required:
        - avgServicePrice
        - avgServiceDurationMinutes
        - effectiveCommissionPercent
        - netProfitPerClient
        - totalCostsMonthly
        - clientsNeededTotal
        - clientsNeededPerEmployee
        - clientsNeededPerDayPerEmployee
        - maxCapacityPerEmployeePerMonth
        - maxCapacityTotalPerMonth
        - utilizationPercent
        - marginPercent
        - isRealistic
        - insights
      properties:
        avgServicePrice:
          type: number
        avgServiceDurationMinutes:
          type: number
        effectiveCommissionPercent:
          type: number
        netProfitPerClient:
          type: number
        totalCostsMonthly:
          type: number
        clientsNeededTotal:
          type: number
        clientsNeededPerEmployee:
          type: number
        clientsNeededPerDayPerEmployee:
          type: number
        maxCapacityPerEmployeePerMonth:
          type: number
        maxCapacityTotalPerMonth:
          type: number
        utilizationPercent:
          type: number
        marginPercent:
          type: number
        isRealistic:
          type: boolean
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
    BreakEvenInput:
      type: object
      required:
        - avgServicePrice
        - commissionModel
        - commissionConfig
        - monthlyOverheadShare
        - estimatedClientsPerDay
      properties:
        avgServicePrice:
          type: number
          minimum: 0
        commissionModel:
          $ref: "#/components/schemas/CommissionModel"
        commissionConfig:
          $ref: "#/components/schemas/CommissionConfig"
        monthlyOverheadShare:
          type: number
          minimum: 0
        estimatedClientsPerDay:
          type: number
          minimum: 0.1
    BreakEvenResult:
      type: object
      required:
        - netProfitPerClient
        - monthlyFixedCostOfHire
        - breakEvenClientsNeeded
        - breakEvenDays
        - breakEvenWeeks
        - breakEvenMonths
        - insights
      properties:
        netProfitPerClient:
          type: number
        monthlyFixedCostOfHire:
          type: number
        breakEvenClientsNeeded:
          type: number
        breakEvenDays:
          type: number
        breakEvenWeeks:
          type: number
        breakEvenMonths:
          type: number
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"

    # ─── HPP ─────────────────────────────────────────────────────────────────
...
    HppInput:
      type: object
      required: [mode, nama]
      properties:
        mode:
          type: string
          enum: [produk, jasa]
        nama:
          type: string
          minLength: 1
        sellingPrice:
          type: ["number", "null"]
          minimum: 0
        # produk fields
        rawMaterials:
          type: array
          items:
            $ref: "#/components/schemas/RawMaterialItem"
          default: []
        directLaborCostPerUnit:
          type: number
          minimum: 0
          default: 0
        monthlyOverhead:
          type: number
          minimum: 0
          default: 0
        estimatedUnitsPerMonth:
          type: number
          minimum: 1
          default: 1
        packagingCostPerUnit:
          type: number
          minimum: 0
          default: 0
        # jasa fields
        consumablesPerSession:
          type: array
          items:
            $ref: "#/components/schemas/RawMaterialItem"
          default: []
        monthlyFixedCosts:
          type: number
          minimum: 0
          default: 0
        estimatedSessionsPerMonth:
          type: number
          minimum: 1
          default: 1
        therapistCommissionPerSession:
          type: number
          minimum: 0
          default: 0
    HppResult:
      type: object
      required: [hpp, insights, breakdown]
      properties:
        hpp:
          type: number
        sellingPrice:
          type: ["number", "null"]
        margin:
          type: ["number", "null"]
        markup:
          type: ["number", "null"]
        breakdown:
          type: object
          additionalProperties:
            type: number
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"

    # ─── BEP Usaha ───────────────────────────────────────────────────────────
    BepUsahaInput:
      type: object
      required: [totalBiayaTetapBulanan, hargaJualPerUnit, hppPerUnit]
      properties:
        totalBiayaTetapBulanan:
          type: number
          minimum: 0
        hargaJualPerUnit:
          type: number
          minimum: 0
        hppPerUnit:
          type: number
          minimum: 0
    BepUsahaResult:
      type: object
      required: [kontribusiMarginPerUnit, unitBepBulanan, revenueBepBulanan, insights]
      properties:
        kontribusiMarginPerUnit:
          type: number
        unitBepBulanan:
          type: number
        revenueBepBulanan:
          type: number
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"

    # ─── Harga Jual ──────────────────────────────────────────────────────────
    HargaJualInput:
      type: object
      required: [mode, hpp]
      properties:
        mode:
          type: string
          enum: [dari_hpp, dari_harga]
        hpp:
          type: number
          minimum: 0
        targetMarginPercent:
          type: ["number", "null"]
          minimum: 0
          maximum: 99.99
        targetMarkupPercent:
          type: ["number", "null"]
          minimum: 0
        hargaJual:
          type: ["number", "null"]
          minimum: 0
    HargaJualResult:
      type: object
      required: [mode, hpp, insights]
      properties:
        mode:
          type: string
        hpp:
          type: number
        hargaDariMargin:
          type: ["number", "null"]
        hargaDariMarkup:
          type: ["number", "null"]
        marginAktual:
          type: ["number", "null"]
        markupAktual:
          type: ["number", "null"]
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"

    # ─── Pajak ───────────────────────────────────────────────────────────────
    PajakInput:
      type: object
      required: [omzetBulanan, skema]
      properties:
        omzetBulanan:
          type: number
          minimum: 0
        skema:
          type: string
          enum: [ppfinal, normal]
        totalBiayaBulanan:
          type: ["number", "null"]
          minimum: 0
    PajakResult:
      type: object
      required: [pajakBulanan, skema, disclaimer, insights]
      properties:
        pajakBulanan:
          type: number
        skema:
          type: string
        disclaimer:
          type: string
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"

    # ─── Ekspansi ────────────────────────────────────────────────────────────
    EkspansiInput:
      type: object
      required: [modalAwal, revenueBulanan, biayaTetapBulanan, hppPercentOfRevenue]
      properties:
        modalAwal:
          type: number
          minimum: 0
        revenueBulanan:
          type: number
          minimum: 0
        biayaTetapBulanan:
          type: number
          minimum: 0
        hppPercentOfRevenue:
          type: number
          minimum: 0
          maximum: 100
    EkspansiResult:
      type: object
      required: [profitBulananProyeksi, paybackPeriodBulan, paybackPeriodTahun, roiTahunan, insights]
      properties:
        profitBulananProyeksi:
          type: number
        paybackPeriodBulan:
          type: number
        paybackPeriodTahun:
          type: number
        roiTahunan:
          type: number
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"

    # ─── Pinjaman ────────────────────────────────────────────────────────────
    PinjamanInput:
      type: object
      required: [plafonPinjaman, sukuBungaTahunan, tenorBulan, metode]
      properties:
        plafonPinjaman:
          type: number
          minimum: 0
        sukuBungaTahunan:
          type: number
          minimum: 0
          maximum: 100
        tenorBulan:
          type: integer
          minimum: 1
          maximum: 360
        metode:
          type: string
          enum: [flat, anuitas]
        profitBulananSaatIni:
          type: ["number", "null"]
          minimum: 0
...
    PinjamanResult:
      type: object
      required: [cicilanBulanan, totalBunga, totalPembayaran, jadwal, insights]
      properties:
        cicilanBulanan:
          type: number
        totalBunga:
          type: number
        totalPembayaran:
          type: number
        profitSetelahCicilan:
          type: ["number", "null"]
        jadwal:
          type: array
          items:
            $ref: "#/components/schemas/JadwalCicilan"
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"

    # ─── Cost Items ──────────────────────────────────────────────────────────
...
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
    AdminAccountSummary:
      type: object
      required:
        - id
        - email
        - businessName
        - tier
        - scenarioCount
        - scenarioLimit
        - exportEnabled
        - benchmarkAccess
        - packageExpiresAt
        - createdAt
      properties:
        id:
          type: integer
        email:
          type: string
        businessName:
          type: ["string", "null"]
        tier:
          $ref: "#/components/schemas/Tier"
        scenarioCount:
          type: integer
        scenarioLimit:
          type: ["integer", "null"]
        exportEnabled:
          type: boolean
        benchmarkAccess:
          type: boolean
        moduleAccess:
          oneOf:
            - $ref: "#/components/schemas/ModuleAccess"
            - type: "null"
        packageExpiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AdminAccountDetail:
      type: object
      required:
        - id
        - email
        - businessName
        - tier
        - scenarioCount
        - scenarioLimit
        - exportEnabled
        - benchmarkAccess
        - packageStartedAt
        - packageExpiresAt
        - createdAt
        - history
      properties:
        id:
          type: integer
        email:
          type: string
        businessName:
          type: ["string", "null"]
        tier:
          $ref: "#/components/schemas/Tier"
        scenarioCount:
          type: integer
        scenarioLimit:
          type: ["integer", "null"]
        exportEnabled:
          type: boolean
        benchmarkAccess:
          type: boolean
        moduleAccess:
          oneOf:
            - $ref: "#/components/schemas/ModuleAccess"
            - type: "null"
        packageStartedAt:
          type: ["string", "null"]
          format: date-time
        packageExpiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        history:
          type: array
          items:
            $ref: "#/components/schemas/AdminHistoryEntry"
    AdminCreateAccountInput:
      type: object
      required: [email, tier]
      properties:
        email:
          type: string
          minLength: 3
        businessName:
          type: ["string", "null"]
        tier:
          $ref: "#/components/schemas/Tier"
        scenarioLimit:
          type: ["integer", "null"]
        exportEnabled:
          type: boolean
        benchmarkAccess:
          type: boolean
        moduleAccess:
          oneOf:
            - $ref: "#/components/schemas/ModuleAccess"
            - type: "null"
        packageStartedAt:
          type: ["string", "null"]
          format: date-time
        packageExpiresAt:
          type: ["string", "null"]
          format: date-time
...